+++ Package.swift
@@ -1,0 +1,6 @@
+import PackageDescription
+
+let package = Package(
+    name: "JsonSwiftson",
+    exclude: ["Graphics", "JsonSwiftson-macOS", "JsonSwiftson-tvOS", "JsonSwiftson-watchOS", "JsonSwiftson.xcodeproj", "JsonSwiftsonAppDemo", "JsonSwiftsonLogicTests", "JsonSwiftsonTests"]
+)