+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 3.0.0 (2016-06-17)
 
 * Converted the library to Swift 3.0.
+* Added support for macOS, tvOS and watchOS.